requirements.txt
biopython
nglview
plotly
numpy
scikit-learn